chore: android ci
