+++ godot/export_presets.cfg
@@ -71,6 +71,7 @@ application/version="1.0"
 application/copyright=""
 application/copyright_localized={}
 application/min_macos_version="10.12"
+application/export_angle=0
 display/high_res=true
 xcode/platform_build="14C18"
 xcode/sdk_version="13.1"
@@ -104,6 +105,7 @@ codesign/entitlements/app_sandbox/files_downloads=0
 codesign/entitlements/app_sandbox/files_pictures=0
 codesign/entitlements/app_sandbox/files_music=0
 codesign/entitlements/app_sandbox/files_movies=0
+codesign/entitlements/app_sandbox/files_user_selected=0
 codesign/entitlements/app_sandbox/helper_executables=[]
 codesign/custom_options=PackedStringArray()
 notarization/notarization=0
@@ -185,6 +187,7 @@ application/product_name=""
 application/file_description=""
 application/copyright=""
 application/trademarks=""
+application/export_angle=0
 ssh_remote_deploy/enabled=false
 ssh_remote_deploy/host="user@host_ip"
 ssh_remote_deploy/port="22"
@@ -213,7 +216,7 @@ custom_features=""
 export_filter="all_resources"
 include_filter="*, *.*"
 exclude_filter=""
-export_path="../../..//Decentraland Godot Rust.apk"
+export_path="../../../dcl-explorer-godot.apk"
 encryption_include_filters=""
 encryption_exclude_filters=""
 encrypt_pck=false
@@ -239,14 +242,14 @@ package/signed=true
 package/app_category=2
 package/retain_data_on_uninstall=false
 package/exclude_from_recents=false
+package/show_in_android_tv=false
+package/show_in_app_library=true
+package/show_as_launcher_app=false
 launcher_icons/main_192x192=""
 launcher_icons/adaptive_foreground_432x432=""
 launcher_icons/adaptive_background_432x432=""
 graphics/opengl_debug=false
 xr_features/xr_mode=0
-xr_features/hand_tracking=0
-xr_features/hand_tracking_frequency=0
-xr_features/passthrough=0
 screen/immersive_mode=true
 screen/support_small=true
 screen/support_normal=true
@@ -404,6 +407,9 @@ permissions/write_sms=false
 permissions/write_social_stream=false
 permissions/write_sync_settings=false
 permissions/write_user_dictionary=false
+xr_features/hand_tracking=0
+xr_features/hand_tracking_frequency=0
+xr_features/passthrough=0
 
 [preset.4]
 
@@ -438,6 +444,7 @@ application/short_version="1.0"
 application/version="1.0"
 application/icon_interpolation=4
 application/launch_screens_interpolation=4
+application/export_project_only=false
 capabilities/access_wifi=false
 capabilities/push_notifications=false
 user_data/accessible_from_files_app=false
