+++ rust/decentraland-godot-lib/Cargo.toml
@@ -59,6 +59,7 @@ chrono = "0.4.31"
 simple-easing = "1.0.1"
 
 [target.'cfg(target_os = "android")'.dependencies]
+ffmpeg-next = { git = "https://github.com/decentraland/rust-ffmpeg/", branch = "audioline-and-mobile-fix", features = ["fix_usize_size_t"] }
 jni = { version = "0.21.1", features = ["invocation"] }
 paranoid-android = "0.2.1"
